Fix several warnings in gdk-pixbuf/io-gif.c
gdk-pixbuf/io-gif.c: In function ‘gif_get_colormap’:
gdk-pixbuf/io-gif.c:277:39: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘unsigned int’ [-Werror=sign-compare]
  while (context->global_colormap_size < context->global_bit_pixel) {
                                       ^
gdk-pixbuf/io-gif.c:286:51: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘unsigned int’ [-Werror=sign-compare]
                 if (context->global_colormap_size == context->background_index) {
                                                   ^~
gdk-pixbuf/io-gif.c: In function ‘gif_get_colormap2’:
gdk-pixbuf/io-gif.c:304:38: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘unsigned int’ [-Werror=sign-compare]
  while (context->frame_colormap_size < context->frame_bit_pixel) {
                                      ^
In file included from /usr/lib/x86_64-linux-gnu/glib-2.0/include/glibconfig.h:9,
                 from /usr/include/glib-2.0/glib/gtypes.h:32,
                 from /usr/include/glib-2.0/glib/galloca.h:32,
                 from /usr/include/glib-2.0/glib.h:30,
                 from /usr/include/glib-2.0/glib/gi18n-lib.h:21,
                 from gdk-pixbuf/io-gif.c:58:
gdk-pixbuf/io-gif.c: In function ‘clip_frame’:
/usr/include/glib-2.0/glib/gmacros.h:336:26: error: comparison of integer expressions of different signedness: ‘unsigned int’ and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
                          ^
gdk-pixbuf/io-gif.c:751:11: note: in expansion of macro ‘MIN’
  *width = MIN (context->width, orig_x + *width) - *x;
           ^~~
/usr/include/glib-2.0/glib/gmacros.h:336:41: error: operand of ?: changes signedness from ‘gint’ {aka ‘int’} to ‘unsigned int’ due to unsignedness of other operand [-Werror=sign-compare]
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
                                         ^~~
gdk-pixbuf/io-gif.c:751:11: note: in expansion of macro ‘MIN’
  *width = MIN (context->width, orig_x + *width) - *x;
           ^~~
/usr/include/glib-2.0/glib/gmacros.h:336:26: error: comparison of integer expressions of different signedness: ‘unsigned int’ and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
                          ^
gdk-pixbuf/io-gif.c:752:12: note: in expansion of macro ‘MIN’
  *height = MIN (context->height, orig_y + *height) - *y;
            ^~~
/usr/include/glib-2.0/glib/gmacros.h:336:41: error: operand of ?: changes signedness from ‘gint’ {aka ‘int’} to ‘unsigned int’ due to unsignedness of other operand [-Werror=sign-compare]
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
                                         ^~~
gdk-pixbuf/io-gif.c:752:12: note: in expansion of macro ‘MIN’
  *height = MIN (context->height, orig_y + *height) - *y;
            ^~~
gdk-pixbuf/io-gif.c: In function ‘gif_get_lzw’:
gdk-pixbuf/io-gif.c:884:78: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
                 context->frame->bg_transparent = (context->gif89.transparent == context->background_index);
                                                                              ^~
In file included from /usr/lib/x86_64-linux-gnu/glib-2.0/include/glibconfig.h:9,
                 from /usr/include/glib-2.0/glib/gtypes.h:32,
                 from /usr/include/glib-2.0/glib/galloca.h:32,
                 from /usr/include/glib-2.0/glib.h:30,
                 from /usr/include/glib-2.0/glib/gi18n-lib.h:21,
                 from gdk-pixbuf/io-gif.c:58:
gdk-pixbuf/io-gif.c: In function ‘gif_prepare_lzw’:
gdk-pixbuf/io-gif.c:1127:43: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’ [-Werror=sign-compare]
         g_assert (context->lzw_clear_code <=
                                           ^~
/usr/include/glib-2.0/glib/gmacros.h:427:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
gdk-pixbuf/io-gif.c:1127:9: note: in expansion of macro ‘g_assert’
         g_assert (context->lzw_clear_code <=
         ^~~~~~~~
gdk-pixbuf/io-gif.c: In function ‘gif_fill_in_lines’:
gdk-pixbuf/io-gif.c:713:6: error: this statement may fall through [-Werror=implicit-fallthrough=]
   if (context->draw_ypos < (context->frame_height - 4)) {
      ^
gdk-pixbuf/io-gif.c:720:2: note: here
  case 1:
  ^~~~
gdk-pixbuf/io-gif.c:723:6: error: this statement may fall through [-Werror=implicit-fallthrough=]
   if (context->draw_ypos < (context->frame_height - 2))
      ^
gdk-pixbuf/io-gif.c:726:2: note: here
  case 2:
  ^~~~

+++ gdk-pixbuf/io-gif.c
@@ -110,7 +110,7 @@ struct _GifContext
         gboolean has_global_cmap;
 
         CMap global_color_map;
-        gint global_colormap_size;
+        gsize global_colormap_size;
         unsigned int global_bit_pixel;
 	unsigned int global_color_resolution;
         unsigned int background_index;
@@ -118,7 +118,7 @@ struct _GifContext
 
         gboolean frame_cmap_active;
         CMap frame_color_map;
-        gint frame_colormap_size;
+        gsize frame_colormap_size;
         unsigned int frame_bit_pixel;
 
 	unsigned int aspect_ratio;
@@ -716,18 +716,24 @@ gif_fill_in_lines (GifContext *context, guchar *dest, guchar v)
 		}
 		/* we don't need a break here.  We draw the outer pixels first, then the
 		 * inner ones, then the innermost ones.  case 0 needs to draw all 3 bands.
-		 * case 1, just the last two, and case 2 just draws the last one*/
-	case 1:
+                 * case 1, just the last two, and case 2 just draws the last one*/
+                /* FALLTHROUGH */
+
+        case 1:
 		if (context->draw_ypos > 2)
 			gif_fill_in_pixels (context, dest, -2, v);
 		if (context->draw_ypos < (context->frame_height - 2))
 			gif_fill_in_pixels (context, dest, 2, v);
 		/* no break as above. */
+                /* FALLTHROUGH */
+
 	case 2:
 		if (context->draw_ypos > 1)
 			gif_fill_in_pixels (context, dest, -1, v);
 		if (context->draw_ypos < (context->frame_height - 1))
 			gif_fill_in_pixels (context, dest, 1, v);
+                /* FALLTHROUGH */
+
 	case 3:
 	default:
 		break;
@@ -742,7 +748,7 @@ clip_frame (GifContext *context,
             gint       *width, 
             gint       *height)
 {
-        gint orig_x, orig_y;
+        guint orig_x, orig_y;
         
         orig_x = *x;
         orig_y = *y;
@@ -881,7 +887,7 @@ gif_get_lzw (GifContext *context)
                         break;
                 }
 
-                context->frame->bg_transparent = (context->gif89.transparent == context->background_index);
+                context->frame->bg_transparent = ((guint) context->gif89.transparent == context->background_index);
                 
                 context->animation->n_frames ++;
                 context->animation->frames = g_list_append (context->animation->frames, context->frame);
@@ -1124,7 +1130,7 @@ gif_prepare_lzw (GifContext *context)
 	context->code_last_byte = 2;
 	context->code_done = FALSE;
 
-        g_assert (context->lzw_clear_code <= 
+        g_assert ((guint) context->lzw_clear_code <=
                   G_N_ELEMENTS (context->lzw_table[0]));
 
 	for (i = 0; i < context->lzw_clear_code; ++i) {
